+++ src/main/java/amerifrance/guideapi/wrapper/AbstractWrapper.java
@@ -12,5 +12,5 @@ public abstract class AbstractWrapper {
 
     public abstract void drawExtras(int mouseX, int mouseY, BaseScreen gui);
 
-    public abstract boolean isMouseOnWrapper(int mouseX, int mouseY);
+    public abstract boolean isMouseOnWrapper(double mouseX, double mouseY);
 }
