+++ src/main/java/api/util/GuiHelper.java
@@ -30,7 +30,7 @@ public class GuiHelper {
      * @param height - Height of the rectangle
      * @return whether or not the mouse is in the rectangle
      */
-    public static boolean isMouseBetween(int mouseX, int mouseY, int x, int y, int width, int height) {
+    public static boolean isMouseBetween(double mouseX, double mouseY, int x, int y, int width, int height) {
         int xSize = x + width;
         int ySize = y + height;
         return (mouseX >= x && mouseX <= xSize && mouseY >= y && mouseY <= ySize);
